Planète sphèrique

+++ MarsRover/Point.cs
@@ -1,10 +1,12 @@
 ﻿namespace MarsRover
 {
-    public record Point(ushort X, ushort Y)
+    public record Point(decimal X, decimal Y)
     {
         public static Point operator +(Point a, Point b)
         {
-            return new Point((ushort) (a.X + b.X), (ushort) (a.Y + b.Y));
+            return new Point((decimal) (a.X + b.X), (decimal) (a.Y + b.Y));
         }
+
+
     }
 }
